fix tests through to test_08_subcmd_extract
This required modification of the extract module, and primersearch module (with
corresponding changes to the target files and inputs) so that primers are
cross-hybridised against their source genomes. Also, the extract tests are fixed
and alignment is now performed when requested. A further try/except clause is used
to skip distance metrics for unaligned amplicons (we can't use AlignIO to read them
in and calculate distances).