fix writing of primersearch amplicon .bed files
The .bed files had previously been reporting the interval region between
the start of the forward primer and the start of the reverse primer, all
taken in the sense of the forward strand. That meant that the "end" of the
region was actually short by the length of the reverse primer. This
change corrects that, and now stores forward and reverse primer start and
end locations in the JSON file.