fix failing test_07 tests
The test_07 and test_08 tests had an issue firstly because the bug fix
to get results to Fera involved changing the command-line API so that
min and max amplicon lengths could be provided as a filter on reported
amplicons. These weren’t incorporated into the test suite at the same
time.