fix primersearch runs for empty primer sets
Primersearch runs where the query primer set was empty would give
exit code 1, which would throw an error in the tests, but not
affect running of the pipeline/workflow. The fix is that, when the
input primer set is empty, we write an empty primersearch output
file, and don't add the job to the list for scheduling.