fix bugs in primersearch parsing, and amplicon extraction
1. There was a bug in the primersearch output parsing where the
last primer in the search output would not be read in correctly.
This affected both classify and extract results.

2. There were problems with calculating amplicon diversity measures
when only one amplicon was produced by a primer set (the distance
matrix could not be calculated). Now, a single-amplicon result
returns zeros.

3. There was a problem when calculating standard deviation when the
amplicon output had only two sequences (a single distance is calculated
which has no variance); this now returns a zero variance.