update tests to delete old output on startup
Previously, tests would write to a predefined output location
(this is necessary to ensure consistency of the JSON output, in
some cases, as the output JSON files refer to speciific paths)
without first clearing those locations.

This could result in tests falsely passing, as prior (passing)
output would be used if an outut file failed to be written.

All test modules and classes now have a startup class method
that defines a specific output directory, and removes the subtree
at that location before running tests.