add primer3_core version check so that inappropriate tests are skipped
At the moment, if primer3_core points to v1, the EMBOSS tests are run,
and primer3 tests are skipped. If primer3_core points to v2+, the
primer3 tests are run, and EMBOSS tests skipped.

In the future, this could be managed more elegantly with a config file
that points to separate primer3_core installations for v1 and v2+