correct for pyline errors in primersearch.py
This was a problematic set of changes. The key chages were for complexity.

One major change is to refactor build_commands(). This was straightforward.

The problematic change was reducing the number of attributes in the
PDPSearchAmplimer class, by using namedtuples for amplimer sequences.
This required modifications in several files of the codebase, including
parsing and comparing JSON files, and in downstream processing of
PrimerSearch output JSON files.